# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: كُتَّاب الرحمن - نظام إدارة دار تحفيظ القرآن الكريم
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Add a new student (auto-generates student code)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a single student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update student info
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CIRCLES (HALAQAT) ───────────────────────────────────────
...
  /vacations:
    get:
      operationId: listVacations
      tags: [vacations]
      summary: List granted vacations
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Vacations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vacation"
    post:
      operationId: createVacation
      tags: [vacations]
      summary: Grant a vacation to a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VacationInput"
      responses:
        "201":
          description: Vacation granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vacation"
  /vacations/{id}:
    delete:
      operationId: deleteVacation
      tags: [vacations]
      summary: Delete a vacation record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── GUARDIANS ───────────────────────────────────────────────
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payment records
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    delete:
      operationId: deletePayment
      tags: [payments]
      summary: Delete a payment record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── EXPENSES ────────────────────────────────────────────────
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, name, studentCode, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        studentCode:
          type: string
        phone:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        enrollmentDate:
          type: ["string", "null"]
        circleId:
          type: ["integer", "null"]
        circleName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    StudentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        enrollmentDate:
          type: string
        circleId:
          type: integer
        notes:
          type: string
    StudentUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        enrollmentDate:
          type: string
        circleId:
          type: ["integer", "null"]
        notes:
          type: string

    # ─── CIRCLE SCHEMAS ───────────────────────────────────────
...
    Vacation:
      type: object
      required: [id, studentId, studentName, startDate, endDate, reason, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    VacationInput:
      type: object
      required: [studentId, startDate, endDate, reason]
      properties:
        studentId:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: string
        notes:
          type: string

    # ─── GUARDIAN SCHEMAS ─────────────────────────────────────
...
    Payment:
      type: object
      required: [id, studentId, studentName, amount, month, paidAt, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        amount:
          type: number
        month:
          type: string
        notes:
          type: ["string", "null"]
        paidAt:
          type: string
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [studentId, amount, month]
      properties:
        studentId:
          type: integer
        amount:
          type: number
        month:
          type: string
        notes:
          type: string
        paidAt:
          type: string

    # ─── EXPENSE SCHEMAS ──────────────────────────────────────
...