# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * كُتَّاب الرحمن - نظام إدارة دار تحفيظ القرآن الكريم
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  studentCode: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  dateOfBirth?: string | null;
  /** @nullable */
  enrollmentDate?: string | null;
  /** @nullable */
  circleId?: number | null;
  /** @nullable */
  circleName?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}